% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot_levels}
\alias{plot_levels}
\title{Plot the predicted antigen presentation levels for spatial transcriptomics data}
\usage{
plot_levels(seurat_object, pathway, mode)
}
\arguments{
\item{seurat_object}{the input seurat object}

\item{pathway}{Select a pathway you want to used to predict the anitgen presentation levels. Choosing from "mhc1","mhc2" or "self_define"}

\item{mode}{The type of antigen presentation levels you want to plot. Choosing from "all", "avg" ,"pre" ,M_1" ,"M_2"..."M_8"}
}
\description{
Plot the predicted antigen presentation levels for spatial transcriptomics data
}
