% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{psaa}
\alias{psaa}
\title{Antigen presentation level prediction using single cell/bulk RNA-seq or spatial transcriptomic data}
\usage{
psaa(seurat_object, pathway, sample_name, protein = FALSE, seed = 123)
}
\arguments{
\item{seurat_object}{Seurat object containing a query data set - antigen presentation level prediction will be applied to this object}

\item{pathway}{Select a pathway you want to used to predict the anitgen presentation levels. Choosing from "mhc1","mhc2" or "self_define"}

\item{sample_name}{The name of your input data}

\item{protein}{Whether to used proteomics data}
}
\value{
seurat object with predicted antigen presentation levels saved in metadata
}
\description{
Antigen presentation level prediction using single cell/bulk RNA-seq or spatial transcriptomic data
}
